added castom hooks
