+++ package.json
@@ -20,7 +20,7 @@
     "react-hook-form": "^7.54.2",
     "react-redux": "^9.2.0",
     "sass": "^1.84.0",
-    "swiper": "^11.2.2"
+    "swiper": "^11.2.3"
   },
   "devDependencies": {
     "@eslint/js": "^9.17.0",
