+++ src/components/ProductCarusel.tsx
@@ -1,35 +1,15 @@
-import { useEffect, useState } from 'react';
 import { Swiper, SwiperSlide } from 'swiper/react';
 import { Navigation, Pagination, Autoplay } from 'swiper/modules';
 import 'swiper/css';
-import axios from 'axios';
-import { Product } from '../types';
 import styles from '../App.module.scss';
+import { fetchProducts } from '../hooks/fetchProducts';
 
 export const ProductCarousel = () => {
-  const [products, setProducts] = useState<Product[]>([]);
-  const [loading, setLoading] = useState(true);
+  const { products, loading, error } = fetchProducts();
 
-  useEffect(() => {
-    const fetchProducts = async () => {
-      try {
-        const response = await axios.get<Product[]>(
-          'https://api.example.com/products'
-        );
-        setProducts(response.data);
-      } catch (error) {
-        console.error('Error fetching products:', error);
-      } finally {
-        setLoading(false);
-      }
-    };
-
-    fetchProducts();
-  }, []);
-
-  if (loading) {
-    return <p>Loading...</p>;
-  }
+  if (loading) return <p>Loading...</p>;
+  if (error) return <p>{error}</p>;
+  if (products.length === 0) return <p>No products found.</p>;
 
   return (
     <Swiper
